[BUGFIX] JS compilation did not recover on errors (#13)
When `gulp serve` is active and an error occured while compiling the JS (e.g. syntax error), the watch task did not continue correctly and only ran the lint task exclusively. So you couldn’t recover from an error because no recompilation was executed.
This is fixed by explicitly calling done() on errors.